requirement.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pillow==9.1.1
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
rich==12.4.4
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
